requirements.txt
Django==3.1.5
factory-boy
pytest
djangorestframework
